#3721: WinCE: Adjust VM for partner-branding - MIDlet install shows "null needs to check for features specific to your phone." and does not start midlet
Implemented a cabinet file searcher into the custom installer to figure
out the partner branding in the file name. The custom installer of the
cab file will set a few registry key value pairs which will be used by
the WinCE VM. If the custom installer cannot find the filename or the
registry does not contain the required x-bw properties, default ones will
be used.